requirements.txt
bson==0.4.8
Django==1.11.5
django-configurations==2.0
django-jet==1.0.6
django-countries==4.6.1
django-widget-tweaks==1.4.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pymongo==3.5.1
pytz==2017.2
